doc-server/requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
Flask-Dance==7.0.0
flask-paginate==2023.10.8
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
psycopg==3.1.12
psycopg2==2.9.9
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
typing_extensions==4.8.0
urllib3==2.0.7
URLObject==2.4.3
Werkzeug==3.0.0
